requirements.txt
python-dotenv
anthropic
numpy
opencv-python
Pillow
shapely
torch
diffusers
transformers
accelerate
gradio
torchvision
matplotlib
